UC-12 refactor: rebase /shared/components and rename folder to /entities
