+++ src/app/store/store.ts
@@ -3,7 +3,7 @@ import { configureStore, combineReducers } from '@reduxjs/toolkit'
 import { setupListeners } from '@reduxjs/toolkit/query/react'
 import { signInReducer } from '@/src/features/auth/signIn/model/signInSlice'
 import { loadState, saveState } from '@/src/shared/lib/localstorage'
-import { viewPostsApi } from '@/src/shared/components/Posts/service/api/viewPost.api'
+import { viewPostsApi } from '@/src/entities/Posts/service/api/viewPost.api'
 
 const rootReducer = combineReducers({
   [baseApi.reducerPath]: baseApi.reducer,
