+++ src/entities/PolicySwitcher/PolicySwitcher.tsx
