+++ src/entities/Posts/service/api/viewPost.api.ts
