+++ src/entities/Posts/service/ui/CircleBlock/CircleBlock.tsx
