+++ src/entities/Posts/service/ui/DescreptionBlock/DescriptionBlock.tsx
