+++ src/entities/Posts/service/ui/SliderArrow/SliderArrow.tsx
