+++ src/pages/home/index.tsx
@@ -1,5 +1,5 @@
-import Posts from '@/src/shared/components/Posts/Posts'
-import { useViewPostsQuery } from '@/src/shared/components/Posts/service/api/viewPost.api'
+import Posts from '@/src/entities/Posts/Posts'
+import { useViewPostsQuery } from '@/src/entities/Posts/service/api/viewPost.api'
 import { getLayoutWithSidebar } from '@/src/widgets/Layout/LayoutWithSidebar'
 import { withAuth } from '@/src/features/private_routes/withAuth'
 import { useEffect, useState } from 'react'
