--- src/shared/components/Posts/service/types/postsTypes.ts
@@ -1,21 +1,0 @@
-type imgType = {
-    imageId: number
-    imageUrl: string
-}
-type likeType = {
-    name: string
-    like: boolean
-    avatar: string
-}
-type comment = {
-    name: string
-    comment: string
-    data: string
-}
-export type PostsProps = {
-    comments: any
-    name: string
-    img: imgType[]
-    postDescription: string | null
-    likes: likeType[]
-}
